i(impr) Dockerfile is now the docker way
The source files are copied on build time and not mounted.
This is the way the Dockerfiles should be written. Unfortunately
due to scopes, the Dockerfile cannot be inside docker, because
the COPY operation does not go above (..) the Dockerfile position.
Alternative to this could be to 'RUN git clone [....]' but this
approach is too rigid: in the way proposed, the files can still
be modified and the image rebuilt.